fixed transfer module

+++ ViewModels/OperationViewModel.cs
@@ -2,7 +2,8 @@
 
 public class OperationViewModel
 {
-    public double Amount { get; set; }
+    public decimal Amount { get; set; }
 
     public string RecipientСardNumber { get; set; }
 }
+
